Fix 1782
