+++ packages/graphql/tests/integration/issues/1221.int.test.ts
@@ -540,164 +540,6 @@ describe("https://github.com/neo4j/graphql/issues/1221", () => {
         });
     });
 
-    test("should only return the single 'chain' and the multiple nested 'chains', three relations deep, using SOME", async () => {
-        const neoGraphql = new Neo4jGraphQL({
-            typeDefs: extendedTypeDefs,
-            driver,
-        });
-        schema = await neoGraphql.getSchema();
-
-        const session = await neo4j.getSession();
-        try {
-            await session.run(`
-                CREATE (:${testNameDetails} { fullName: "MHA" })<-[:HAS_NAME { current: true }]-(:${testMasterData} { current: true, id: "123" })<-[:ARCHITECTURE { current: true }]-(:${testSeries} { current: true, id: "321" })<-[:MAIN { current: true }]-(:${testMain} { current: true, id: "1321" })
-                CREATE (s:${testSeries} { current: true, id: "421" })
-                CREATE (:${testNameDetails} { fullName: "MHA" })<-[:HAS_NAME { current: true }]-(:${testMasterData} { current: true, id: "123" })<-[:ARCHITECTURE { current: true }]-(s)<-[:MAIN { current: true }]-(:${testMain} { current: true, id: "1322" })
-                CREATE (:${testNameDetails} { fullName: "MHA" })<-[:HAS_NAME { current: true }]-(:${testMasterData} { current: true, id: "1123" })<-[:ARCHITECTURE { current: true }]-(s)
-               
-                // For verification purpose, this should be filtered out by the where clause:
-                CREATE (:${testNameDetails} { fullName: "MHBB" })<-[:HAS_NAME { current: true }]-(:${testMasterData} { current: true, id: "523" })<-[:ARCHITECTURE { current: true }]-(:${testSeries} { current: true, id: "621" })<-[:MAIN { current: true }]-(:${testMain} { current: true, id: "1621" })
-            `);
-        } finally {
-            await session.close();
-        }
-
-        const query = `
-                query (
-                    $where: ${testMain}Where = { current: true }
-                    $connectionWhere: RelationPropsWhere = { current: true }
-                ) {
-                    ${testMain.plural}(where: $where) {
-                        id
-                        mainConnection(where: { edge: $connectionWhere }) {
-                            edges {
-                                node {
-                                    architectureConnection(where: { edge: $connectionWhere }) {
-                                        edges {
-                                            node {
-                                                nameDetailsConnection(where: { edge: $connectionWhere }) {
-                                                    edges {
-                                                        node {
-                                                            fullName
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-        `;
-
-        const variableValues = {
-            where: {
-                current: true,
-                mainConnection_SINGLE: {
-                    node: {
-                        architectureConnection_SOME: {
-                            node: {
-                                nameDetailsConnection: {
-                                    node: {
-                                        fullName: "MHA",
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-            connectionWhere: {
-                current: true,
-            },
-        };
-
-        const res = await graphql({
-            schema,
-            source: query,
-            variableValues,
-            contextValue: neo4j.getContextValues(),
-        });
-
-        expect(res.errors).toBeUndefined();
-
-        expect(res.data).toEqual({
-            [testMain.plural]: [
-                {
-                    mainConnection: {
-                        edges: [
-                            {
-                                node: {
-                                    architectureConnection: {
-                                        edges: [
-                                            {
-                                                node: {
-                                                    nameDetailsConnection: {
-                                                        edges: [
-                                                            {
-                                                                node: {
-                                                                    fullName: "MHA",
-                                                                },
-                                                            },
-                                                        ],
-                                                    },
-                                                },
-                                            },
-                                        ],
-                                    },
-                                },
-                            },
-                        ],
-                    },
-                    id: "1321",
-                },
-                {
-                    mainConnection: {
-                        edges: [
-                            {
-                                node: {
-                                    architectureConnection: {
-                                        edges: [
-                                            {
-                                                node: {
-                                                    nameDetailsConnection: {
-                                                        edges: [
-                                                            {
-                                                                node: {
-                                                                    fullName: "MHA",
-                                                                },
-                                                            },
-                                                        ],
-                                                    },
-                                                },
-                                            },
-                                            {
-                                                node: {
-                                                    nameDetailsConnection: {
-                                                        edges: [
-                                                            {
-                                                                node: {
-                                                                    fullName: "MHA",
-                                                                },
-                                                            },
-                                                        ],
-                                                    },
-                                                },
-                                            },
-                                        ],
-                                    },
-                                },
-                            },
-                        ],
-                    },
-                    id: "1322",
-                },
-            ],
-        });
-    });
-
     test("should return as two single 'chains', three relations deep", async () => {
         const neoGraphql = new Neo4jGraphQL({
             typeDefs: extendedTypeDefs,
